requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
altair>=5.0.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
xlrd>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
